Improved IXO_STRUCTDEF
Removed typedef from IXO_STRUCTDEF, which was dependent on the line number, this could lead to strange errors, if another IXO_STRUCTDEF was invoked on the same linenumber in another file. It is done by passing an additional tuple to IXO_APPLY_EACH, which represents the capture/closure of the macro, that is applied to each tuple in the variadic parameter list.